ll -> int

+++ Math/PowMod.cpp
@@ -1,4 +1,4 @@
-long long powmod(long long base, long long power, long long mod) {
+long long powmod(long long base, long long power, int mod) {
   long long ans = 1;
   while (power > 0) {
     if (power & 1) { ans = ans * base % mod; }
